moved to double, improved voxelizer

+++ RTSPathingLib/Source/RTSPathingLib/Body.h
@@ -4,19 +4,19 @@
 
 namespace RTSPathingLib {
   struct Body {
-    Body(const glm::vec2& pos, size_t category_, size_t size_ = 1) {
+    Body(const glm::dvec2& pos, size_t category_, size_t size_ = 1) {
       position = pos;
       category = category_;
       size     = size_;
     }
 
     Body() {
-      position = glm::vec2(0, 0);
+      position = glm::dvec2(0, 0);
       category = 0;
       size     = 1;
     }
 
-    glm::vec2 position;
+    glm::dvec2 position;
     size_t    category;
     size_t    size    ;
   };
