+++ RTSPathingLib/Source/RTSPathingLib/BodyUtil.h
@@ -10,6 +10,6 @@ namespace RTSPathingLib {
   public:
     static Body              transform  (const Body& b, const glm::mat4& transformation);
     static std::vector<Body> transform  (const std::vector<Body>& b, const glm::mat4& transformation);
-    static glm::vec2         getCenter  (const std::vector<Body>& b);
+    static glm::dvec2         getCenter  (const std::vector<Body>& b);
   };
 }
