+++ RTSPathingLib/Source/RTSPathingLib/Formation/FormationShape/ArcFormationShape.h
@@ -10,17 +10,17 @@ namespace RTSPathingLib {
 
     virtual std::string getType() const override;
 
-    virtual std::vector<glm::vec2> getPolygon() const        override;
+    virtual std::vector<glm::dvec2> getPolygon() const        override;
     virtual size_t getNumberOfInterfacePoints() const        override;
-    virtual glm::vec2 getInterfacePoint(size_t number) const override; //center, border+1/36, border+2/36,...
-    virtual glm::vec2 getInterfaceNormal(size_t number) const override;
-    virtual float getInterfaceWidth(size_t number) const;
+    virtual glm::dvec2 getInterfacePoint(size_t number) const override; //center, border+1/36, border+2/36,...
+    virtual glm::dvec2 getInterfaceNormal(size_t number) const override;
+    virtual double getInterfaceWidth(size_t number) const;
 
-    float getArcAngle() const;
-    void setArcAngle(float valueRadian);
+    double getArcAngle() const;
+    void setArcAngle(double valueRadian);
 
   private:
-    float arcAngleRadian = 0.0f; // for e.g half circles. 0 means full circle, pi means half circle, 2pi means nothing
+    double arcAngleRadian = 0.0; // for e.g half circles. 0 means full circle, pi means half circle, 2pi means nothing
 
     const size_t numberPolygons = 36;
   };
