+++ RTSPathingLib/Source/RTSPathingLib/Formation/FormationShape/RectangleFormationShape.h
@@ -10,11 +10,11 @@ namespace RTSPathingLib {
 
     virtual std::string getType() const override;
 
-    virtual std::vector<glm::vec2> getPolygon() const        override;
+    virtual std::vector<glm::dvec2> getPolygon() const        override;
     virtual size_t getNumberOfInterfacePoints() const        override;
-    virtual glm::vec2 getInterfacePoint(size_t number) const override; //center, edge0 start, edge0 center, edge1 start,... edge3 center
-    virtual glm::vec2 getInterfaceNormal(size_t number) const override;
-    virtual float getInterfaceWidth(size_t number) const;
+    virtual glm::dvec2 getInterfacePoint(size_t number) const override; //center, edge0 start, edge0 center, edge1 start,... edge3 center
+    virtual glm::dvec2 getInterfaceNormal(size_t number) const override;
+    virtual double getInterfaceWidth(size_t number) const;
 
 
   private:
