+++ RTSPathingLib/Source/RTSPathingLib/Formation/FormationShape/TriangleFormationShape.cpp
@@ -19,28 +19,28 @@ namespace RTSPathingLib {
     return 7;
   }
 
-  glm::vec2 TriangleFormationShape::getInterfacePoint(size_t number) const {
+  glm::dvec2 TriangleFormationShape::getInterfacePoint(size_t number) const {
     if (number == 0)
-      return glm::vec2(0, 0);
+      return glm::dvec2(0, 0);
     size_t edge = ((number - 1) / 2) % 3;
     bool center = (number - 1) % 2 == 0;
     auto poly = getPolygon();
-    glm::vec2 result = poly[edge];
+    glm::dvec2 result = poly[edge];
     if (center)
-      result += (poly[(edge + 1) % 3] - poly[edge]) / 2.0f;
+      result += (poly[(edge + 1) % 3] - poly[edge]) / 2.0;
     return result;
   }
 
-  std::vector<glm::vec2> TriangleFormationShape::getPolygon() const {
-    glm::vec2 size = getBaseSize() * 0.95f;
-    return std::vector<glm::vec2>{
-      glm::vec2(size.x* std::sin((glm::pi<float>() * 2.0f) * (0.0000 + 0.5)), size.y* std::cos((glm::pi<float>() * 2.0f) * (0.0000 + 0.5))),
-      glm::vec2(size.x* std::sin((glm::pi<float>() * 2.0f) * (0.3333 + 0.5)), size.y* std::cos((glm::pi<float>() * 2.0f) * (0.3333 + 0.5))),
-      glm::vec2(size.x* std::sin((glm::pi<float>() * 2.0f) * (0.6666 + 0.5)), size.y* std::cos((glm::pi<float>() * 2.0f) * (0.6666 + 0.5))),
+  std::vector<glm::dvec2> TriangleFormationShape::getPolygon() const {
+    glm::dvec2 size = getBaseSize() * 0.95;
+    return std::vector<glm::dvec2>{
+      glm::dvec2(size.x* std::sin((glm::pi<double>() * 2.0) * (0.0000 + 0.5)), size.y* std::cos((glm::pi<double>() * 2.0) * (0.0000 + 0.5))),
+      glm::dvec2(size.x* std::sin((glm::pi<double>() * 2.0) * (0.3333 + 0.5)), size.y* std::cos((glm::pi<double>() * 2.0) * (0.3333 + 0.5))),
+      glm::dvec2(size.x* std::sin((glm::pi<double>() * 2.0) * (0.6666 + 0.5)), size.y* std::cos((glm::pi<double>() * 2.0) * (0.6666 + 0.5))),
     };
   }
 
-  float TriangleFormationShape::getInterfaceWidth(size_t number) const {
+  double TriangleFormationShape::getInterfaceWidth(size_t number) const {
     if (number == 0)
       return 0;
     size_t edge = ((number - 1) / 2) % 3;
@@ -48,22 +48,22 @@ namespace RTSPathingLib {
     if (!center)
       return 0;
     auto poly = getPolygon();
-    float result = glm::distance(poly[edge], poly[(edge + 1) % 3]);
+    double result = glm::distance(poly[edge], poly[(edge + 1) % 3]);
     return result;
   }
 
-  glm::vec2 TriangleFormationShape::getInterfaceNormal(size_t number) const {
+  glm::dvec2 TriangleFormationShape::getInterfaceNormal(size_t number) const {
     if (number == 0)
-      return glm::vec2(0, 1);
+      return glm::dvec2(0, 1);
     size_t edge = (number - 1) % 3;
     bool center = (number - 1) % 2 == 0;
     if (!center)
-      return glm::vec2(0, 1);
+      return glm::dvec2(0, 1);
     auto poly = getPolygon();
-    glm::vec2 result = glm::normalize(poly[(edge + 1) % 3] - poly[edge]);
-    result = glm::vec2(-result.y, result.x);
+    glm::dvec2 result = glm::normalize(poly[(edge + 1) % 3] - poly[edge]);
+    result = glm::dvec2(-result.y, result.x);
     if (glm::length(result) < 1e-5)
-      return glm::vec2(0, 1);
+      return glm::dvec2(0, 1);
     return result;
   }
 }
