+++ RTSPathingLib/Source/RTSPathingLib/Util/HexagonGrid/HexGridSvg.h
@@ -8,9 +8,9 @@ namespace RTSPathingLib {
 
   class HexGridSvg {
   public:
-    static std::vector<svg> write(const TopFlatHexGrid<bool>& grid, float scale = 1);
+    static std::vector<svg> write(const TopFlatHexGrid<bool>& grid, double scale = 1);
 
   private:
-    static std::vector<glm::vec2> getHexagon(const glm::vec2& offset, float radius);
+    static std::vector<glm::dvec2> getHexagon(const glm::dvec2& offset, double radius);
   };
 }
