+++ RTSPathingLib/Source/RTSPathingLib/Util/RTree2D.cpp
@@ -7,8 +7,8 @@
 #include <boost/geometry/index/rtree.hpp>
 
 #include "glmBoostGeometry.h"
-BOOST_GEOMETRY_REGISTER_POINT_2D(glm::vec2, float, cs::cartesian, x, y)
-BOOST_GEOMETRY_REGISTER_BOX(RTSPathingLib::glm2dBox, glm::vec2, lowLeft, upRight)
+BOOST_GEOMETRY_REGISTER_POINT_2D(glm::dvec2, double, cs::cartesian, x, y)
+BOOST_GEOMETRY_REGISTER_BOX(RTSPathingLib::glm2dBox, glm::dvec2, lowLeft, upRight)
 
 namespace RTSPathingLib {
   //https://stackoverflow.com/questions/34951426/cannot-remove-element-from-boostgeometryindexrtree-by-its-index
@@ -23,13 +23,13 @@ namespace RTSPathingLib {
 
   class RTree2D::RTreeImpl {
   public:
-    std::unique_ptr<boost::geometry::index::rtree<std::pair<glm::vec2, size_t>, boost::geometry::index::dynamic_rstar>> tree = nullptr;
-    std::map<size_t, std::pair<glm::vec2, size_t>> data;
+    std::unique_ptr<boost::geometry::index::rtree<std::pair<glm::dvec2, size_t>, boost::geometry::index::dynamic_rstar>> tree = nullptr;
+    std::map<size_t, std::pair<glm::dvec2, size_t>> data;
   };
 
-  RTree2D::RTree2D(const std::vector<std::pair<glm::vec2, size_t>>& bulkInput) {
+  RTree2D::RTree2D(const std::vector<std::pair<glm::dvec2, size_t>>& bulkInput) {
     p = std::make_unique<RTree2D::RTreeImpl>();
-    p->tree = std::make_unique<boost::geometry::index::rtree<std::pair<glm::vec2, size_t>, boost::geometry::index::dynamic_rstar>>(bulkInput, boost::geometry::index::dynamic_rstar(bulkInput.size()));
+    p->tree = std::make_unique<boost::geometry::index::rtree<std::pair<glm::dvec2, size_t>, boost::geometry::index::dynamic_rstar>>(bulkInput, boost::geometry::index::dynamic_rstar(bulkInput.size()));
 
     for (const auto& x : bulkInput)
       p->data[x.second] = x;
@@ -39,8 +39,8 @@ namespace RTSPathingLib {
 
   }
 
-  std::vector<size_t> RTree2D::boxQuery(const glm::vec2& lowLeftBox, const glm::vec2& upRightBox) {
-    std::vector<std::pair<glm::vec2, size_t>> preResult;
+  std::vector<size_t> RTree2D::boxQuery(const glm::dvec2& lowLeftBox, const glm::dvec2& upRightBox) {
+    std::vector<std::pair<glm::dvec2, size_t>> preResult;
     glm2dBox box1(lowLeftBox, upRightBox);
     p->tree->query(boost::geometry::index::covered_by(box1), std::back_inserter(preResult));
 
@@ -51,9 +51,9 @@ namespace RTSPathingLib {
     return result;
   }
 
-  std::vector<size_t> RTree2D::nearestNeighbour(const glm::vec2 position, size_t amount) {
+  std::vector<size_t> RTree2D::nearestNeighbour(const glm::dvec2 position, size_t amount) {
     //https://gist.github.com/warrenrentlytics/c9a1836a40d4fcbba28a7e29357dad7d
-    std::vector<std::pair<glm::vec2, size_t>> preResult;
+    std::vector<std::pair<glm::dvec2, size_t>> preResult;
     p->tree->query(boost::geometry::index::nearest(position, amount), std::back_inserter(preResult));
 
     std::vector<size_t> result;
