+++ RTSPathingLib/Source/RTSPathingLib/Util/RTree2D.h
@@ -8,11 +8,11 @@
 namespace RTSPathingLib {
   class RTree2D {
   public:
-    RTree2D(const std::vector<std::pair<glm::vec2, size_t>>& bulkInput);
+    RTree2D(const std::vector<std::pair<glm::dvec2, size_t>>& bulkInput);
     virtual ~RTree2D();
 
-    std::vector<size_t> boxQuery(const glm::vec2& lowLeftBox, const glm::vec2& upRightBox);
-    std::vector<size_t> nearestNeighbour(const glm::vec2 position, size_t amount = 1);
+    std::vector<size_t> boxQuery(const glm::dvec2& lowLeftBox, const glm::dvec2& upRightBox);
+    std::vector<size_t> nearestNeighbour(const glm::dvec2 position, size_t amount = 1);
 
     void remove(size_t index);
