+++ RTSPathingLib/Source/RTSPathingLib/Util/RectangleGrid/RectangleGrid.h
@@ -9,6 +9,6 @@ namespace RTSPathingLib {
   public:
     glm::ivec2 dimension = glm::ivec2(1,1);
     std::vector<T> data;
-    glm::vec2 offset = glm::vec2(0, 0);
+    glm::dvec2 offset = glm::dvec2(0, 0);
   };
 }
