+++ RTSPathingLib/Source/RTSPathingLib/Util/RectangleGrid/RectangleGridSvg.h
@@ -8,9 +8,9 @@ namespace RTSPathingLib {
 
   class RectangleGridSvg {
   public:
-    static std::vector<svg> write(const RectangleGrid<bool>& grid, float scale = 1);
+    static std::vector<svg> write(const RectangleGrid<bool>& grid, double scale = 1);
 
   private:
-    static std::vector<glm::vec2> getRectangle(const glm::vec2& offset, float scale);
+    static std::vector<glm::dvec2> getRectangle(const glm::dvec2& offset, double scale);
   };
 }
